+++ functions_test.py
@@ -1,14 +1,9 @@
 import logging
-# from interact_onos import start_onos_docker, toggle_fwd, create_host_intents, get_intents, clear_intent, clear_all_intents
 from interact_onos import *
 from interact_mininet import MininetProcess
 from time import sleep
-import numpy as np
-import matplotlib.pyplot as plt 
 import datetime
 import os
-import statistics
-import statistics
 
 def clear_log_file(log_file_path):
     try:
@@ -310,351 +305,6 @@ def dynamic_paths_test(testcase, topo="linear,3,2"):
 
     logging.info(f"{test_name} test done")
 
-def fault_tolerance_test(testcase, topo="TC"):
-    test_name = "Fault Tolerance"
-    testcase_success = False
-    start_time = datetime.datetime.now()
-    start_time_str = start_time.strftime("%Y-%m-%d-%H%M")
-    sleep_time = 3
-    iperf_time = 20 # 20s iperf test
-    trial_runs = 20 # 20 for number of trials
-    failover_time = iperf_time/2
-
-    logging.info(f"Starting {test_name} test at {start_time_str}")
-    logging.info(f"iperf_time: {iperf_time}s, trial_runs: {trial_runs}")
-
-    if testcase == "fwd":
-        testcase_name = "Fault Tolerance via Reactive Fwd"
-    elif testcase == "intent":
-        testcase_name = "Fault Tolerance via Host Intents"
-    else:
-        logging.error(f"Invalid testcase: {testcase}")
-        return
-    
-    try:
-        # Starting actions
-        output = []
-        baseline_capacity_data = []
-        failover_capacity_data = []
-        baseline_throughput_data = []
-        failover_throughput_data = []
-        capacity_diff_data = []
-        throughput_diff_data = []
-
-        output = []
-        baseline_capacity_data = []
-        failover_capacity_data = []
-        baseline_throughput_data = []
-        failover_throughput_data = []
-        capacity_diff_data = []
-        throughput_diff_data = []
-
-        start_onos_docker()
-        mininet_process = MininetProcess(topo)
-        mininet_process.start_mininet()
-
-        logging.info(f"Test Case: {testcase_name}")
-
-        if testcase == "fwd":
-            toggle_fwd("activate")
-        elif testcase == "intent":
-            toggle_fwd("activate")
-            sleep(sleep_time)
-            mininet_process.send_command("pingall")
-            mininet_process.read_stderr("*** Results")
-            toggle_fwd("deactivate")
-            clear_all_intents()
-            create_host_intents(get_mac_addresses())
-        else:
-            logging.error(f"Invalid testcase: {testcase}")
-            return
-
-        # Start iperf server
-        sleep(sleep_time)
-        mininet_process.send_command("h1 iperf -s &", check_stdout=True)
-
-        for i in range(trial_runs):
-            paths_1 = None
-            paths_2 = None
-            baseline_capacity = None
-            failover_capacity = None
-            baseline_throughput = None
-            failover_throughput = None
-            capacity_diff = None
-            throughput_diff = None
-
-            sleep(sleep_time)
-
-            # Baseline iperf test
-            mininet_process.send_command(f"h6 iperf -c 10.0.10.1 -t {iperf_time} -i 1 -f mM > iperf/iperf_{testcase}_baseline.log &", check_stdout=True)
-            sleep(iperf_time)
-            logging.info("Baseline iperf output")
-            baseline_output = mininet_process.read_logfile(f"iperf/iperf_{testcase}_baseline.log")
-
-            # Failover iperf test
-            mininet_process.send_command(f"h6 iperf -c 10.0.10.1 -t {iperf_time} -i 1 -f mM > iperf/iperf_{testcase}_failover.log &", check_stdout=True)
-            sleep(failover_time)
-            mac_list = get_mac_addresses()
-            paths_1 = get_all_paths(mac_list)
-            logging.info(f"Failover at time {failover_time}s")
-            link_failover_cmd = "link s1 s3 down"
-            mininet_process.send_command(link_failover_cmd, check_stdout=True)
-            paths_2 = get_all_paths(mac_list)
-            logging.info("Failover iperf output")
-            failover_output = mininet_process.read_logfile(f"iperf/iperf_{testcase}_failover.log")
-
-            # Compare paths
-            if paths_1 !=  paths_2:
-                logging.info(f"Paths are different after {link_failover_cmd} command")
-                for path_list in paths_2:
-                    if path_list not in paths_1:
-                        if path_list:
-                            logging.info("Added Paths: " + str(path_list))
-                for path_list in paths_1:
-                    if path_list not in paths_2:
-                        if path_list:
-                            logging.info("Removed Paths: " + str(path_list))
-            else:
-                logging.info("Paths are the same")  
-                logging.error(f"Test Case: {testcase_name} Trial {i+1} Fail")
-                return False 
-
-            # Parse iperf output
-            for lines in baseline_output.split("  "):
-                if "MBytes" in lines:
-                    baseline_capacity = float(lines.split(" ")[0])
-                if "Mbits/sec" in lines:
-                    baseline_throughput = float(lines.split(" ")[0])
-            for lines in failover_output.split("  "):
-                if "MBytes" in lines:
-                    failover_capacity = float(lines.split(" ")[0])
-                if "Mbits/sec" in lines:
-                    failover_throughput = float(lines.split(" ")[0])
-            
-            # Log difference between baseline and failover
-            logging.info(f"Baseline Capacity: {baseline_capacity} Mbytes, Baseline Throughput: {baseline_throughput} Mbits/sec")
-            logging.info(f"Failover Capacity: {failover_capacity} Mbytes, Failover Throughput: {failover_throughput} Mbits/sec")
-            capacity_diff = round(float(baseline_capacity) - float(failover_capacity), 2)
-            throughput_diff = round(float(baseline_throughput) - float(failover_throughput), 2)
-            logging.info(f"Capacity Difference: {capacity_diff} Mbytes, Throughput Difference: {throughput_diff} Mbits/sec")
-
-            # Logging success/failure
-            if paths_1 != paths_2:
-                testcase_success = True
-            if testcase_success:
-                logging.info(f"{testcase_name} Test Trial {i+1} Success")
-                timestring = datetime.datetime.now().strftime("%H:%M:%S")
-                print(f"{testcase_name} Test Trial {i+1} Success, time: {timestring}")
-                # print(f"Baseline Capacity: {baseline_capacity} Mbytes, Baseline Throughput: {baseline_throughput} Mbits/sec")
-                # print(f"Failover Capacity: {failover_capacity} Mbytes, Failover Throughput: {failover_throughput} Mbits/sec")
-                # print(f"Capacity Difference: {capacity_diff} Mbytes, Throughput Difference: {throughput_diff} Mbits/sec")
-                baseline_capacity_data.append(baseline_capacity)
-                failover_capacity_data.append(failover_capacity)
-                baseline_throughput_data.append(baseline_throughput)
-                failover_throughput_data.append(failover_throughput)
-                capacity_diff_data.append(capacity_diff)
-                throughput_diff_data.append(throughput_diff)
-            else:
-                logging.error(f"{testcase_name} Test Fail")
-                print(f"{testcase_name} Test Fail")
-                return False
-            
-            # Reset link
-            logging.info("Resetting link")
-            link_reset_cmd1 = "link s1 s3 up"
-            mininet_process.send_command(link_reset_cmd1, check_stdout=True)
-            link_reset_cmd2 = "link s1 s2 down"
-            mininet_process.send_command(link_reset_cmd2, check_stdout=True)
-            link_reset_cmd3 = "link s1 s2 up"
-            mininet_process.send_command(link_reset_cmd3, check_stdout=True)
-            sleep(1)
-        
-        # Closing actions
-        output = (baseline_capacity_data, failover_capacity_data, baseline_throughput_data, failover_throughput_data, capacity_diff_data, throughput_diff_data)
-        mininet_process.process.stdin.close()
-        mininet_process.read_stderr("Done")
-        logging.info(f"Output: {output}")
-
-        baseline_capacity_data = np.array(baseline_capacity_data, dtype=float)
-        failover_capacity_data = np.array(failover_capacity_data, dtype=float)
-        baseline_throughput_data = np.array(baseline_throughput_data, dtype=float)
-        failover_throughput_data = np.array(failover_throughput_data, dtype=float)
-        capacity_diff_data = np.array(capacity_diff_data, dtype=float)
-        throughput_diff_data = np.array(throughput_diff_data, dtype=float)
-
-        # boxplot
-        fig, ax = plt.subplots(1, 2)
-        fig.set_size_inches(10, 5)
-        ax[0].boxplot([baseline_capacity_data, failover_capacity_data], widths=0.4, labels=["Baseline", "Single Link Failover"])
-        ax[0].set_title("Capacity of 20s iPerf Test")
-        # ticks = np.arange(0, 32, 2)
-        M = max(max(baseline_capacity_data), max(failover_capacity_data))
-        M = M + 10 - M % 10
-        ticks = np.arange(0, M+2, 2)
-        ax[0].set_yticks(ticks)
-        ax[0].set_ylabel("MBytes")
-        failover_median = statistics.median(failover_capacity_data)
-        baseline_median = statistics.median(baseline_capacity_data)
-        diff_median = (failover_median - baseline_median) / baseline_median * 100
-        ax[0].text(1.6, failover_median, f"{'+' if diff_median >= 0 else '-'}{diff_median:.2f}% ", ha='center', va='center', color='red')
-        ax[0].text(1.5, 0, "Median % Difference from Baseline", ha='center', va='bottom', color='red')
-
-        ax[1].boxplot([baseline_throughput_data, failover_throughput_data], widths=0.4, labels=["Baseline", "Single Link Failover"])
-        ax[1].set_title("Throughput of 20s iPerf Test")
-        ax[1].set_ylabel("Mbits/sec")
-        # ticks = np.arange(0, 11, 1)
-        M = max(max(baseline_throughput_data), max(failover_throughput_data))
-        M = M + 10 - M % 10
-        ticks = np.arange(0, M+1, 1)
-        ax[1].set_yticks(ticks)
-        failover_median = statistics.median(failover_throughput_data)
-        baseline_median = statistics.median(baseline_throughput_data)
-        diff_median = (failover_median - baseline_median) / baseline_median * 100
-        ax[1].text(1.6, failover_median, f"{'+' if diff_median >= 0 else '-'}{diff_median:.2f}% ", ha='center', va='center', color='red')
-        ax[1].text(1.5, 0, "Median % Difference from Baseline", ha='center', va='bottom', color='red')                      
-
-        # plt.show()
-        fig.savefig(f"plots/{start_time_str}_{testcase}_{iperf_time}sx{trial_runs}trials_boxplot.png")
-        logging.info(f"{testcase} boxplot saved")
-
-        elapsed_time = (datetime.datetime.now() - start_time)
-        logging.info(f"{test_name} test done, elapsed time: {elapsed_time}")
-        print(f"Elapsed time: {elapsed_time}")
-        return output
-
-    except Exception as e:
-        logging.error(f"Error during {test_name} test: {e}")
-
-def iperf3_test(testcase, topo="TC", iperf_time=10, trial_runs=2, parallel_connections=1):
-    test_name = "iPerf3"
-    testcase_success = False
-    start_time = datetime.datetime.now()
-    start_time_str = start_time.strftime("%Y-%m-%d-%H%M")
-    sleep_time = 3
-    failover_time = iperf_time/2
-    baseline_output = []
-    failover_output = []
-    baseline_log_path = f"iperf/baseline.log"
-    failover_log_path = f"iperf/failover.log"
-
-    logging.info(f"Starting {test_name} test at {start_time_str}")
-    logging.info(f"iperf_time: {iperf_time}s, trial_runs: {trial_runs}, parallel_connections: {parallel_connections}")
-    print(f"Starting {test_name} test with {trial_runs} trials: iperf3 -t {iperf_time}s -P {parallel_connections}")
-
-    if testcase == "fwd":
-        testcase_name = "Fault Tolerance via Reactive Fwd"
-    elif testcase == "intent":
-        testcase_name = "Fault Tolerance via Host Intents"
-    else:
-        logging.error(f"Invalid testcase: {testcase}")
-        return
-    
-    try:
-        # Starting actions
-        start_onos_docker()
-        logging.info(f"Test Case: {testcase_name}")
-        
-        for i in range(trial_runs):
-            paths_1 = None
-            paths_2 = None
-
-            mininet_process = MininetProcess(topo)
-            mininet_process.start_mininet()
-            if testcase == "fwd":
-                toggle_fwd("activate")
-            elif testcase == "intent":
-                toggle_fwd("activate")
-                sleep(sleep_time)
-                mininet_process.send_command("pingall")
-                mininet_process.read_stderr("*** Results")
-                toggle_fwd("deactivate")
-                clear_all_intents()
-                create_host_intents(get_mac_addresses())
-            else:
-                logging.error(f"Invalid testcase: {testcase}")
-                return
-
-            # Start iperf server
-            sleep(sleep_time)
-            mininet_process.send_command("h1 iperf3 -s &", check_stdout=True)
-            sleep(sleep_time)
-
-            # Baseline iperf test
-            mininet_process.send_command(f"h6 iperf3 -c 10.0.10.1 -t {iperf_time} -i 1 -f mM -P {parallel_connections} > {baseline_log_path} &", check_stdout=True)
-            sleep(iperf_time)
-            logging.info("Baseline iperf output")
-            baseline_output.extend(mininet_process.read_iperf3_logfile(baseline_log_path))
-
-            # Failover iperf test
-            sleep(sleep_time)
-            mininet_process.send_command(f"h6 iperf3 -c 10.0.10.1 -t {iperf_time} -i 1 -f mM -P {parallel_connections} > {failover_log_path} &", check_stdout=True)
-            mac_list = get_mac_addresses()
-            paths_1 = get_all_paths(mac_list)
-            sleep(failover_time)
-            logging.info(f"Failover at time {failover_time}s")
-            link_failover_cmd = "link s1 s3 down"
-            mininet_process.send_command(link_failover_cmd, check_stdout=True)
-            sleep(sleep_time)
-            paths_2 = get_all_paths(mac_list)
-            sleep(failover_time)
-            logging.info("Failover iperf output")
-            failover_output.extend(mininet_process.read_iperf3_logfile(failover_log_path))
-
-            # Compare paths
-            if paths_1 !=  paths_2:
-                logging.info(f"Paths are different after {link_failover_cmd} command")
-                for path_list in paths_2:
-                    if path_list not in paths_1:
-                        if path_list:
-                            logging.info("Added Paths: " + str(path_list))
-                for path_list in paths_1:
-                    if path_list not in paths_2:
-                        if path_list:
-                            logging.info("Removed Paths: " + str(path_list))
-            else:
-                logging.info("Paths are the same")  
-                logging.error(f"Test Case: {testcase_name} Trial {i+1} Fail")
-                return False 
-
-            # Logging success/failure
-            if paths_1 != paths_2:
-                testcase_success = True
-            if testcase_success:
-                logging.info(f"{testcase_name} Test Trial {i+1} Success")
-                timestring = datetime.datetime.now().strftime("%H:%M:%S")
-                print(f"{testcase_name} Test Trial {i+1} Success, time: {timestring}")
-            else:
-                logging.error(f"{testcase_name} Test Fail")
-                print(f"{testcase_name} Test Fail")
-                return False
-
-            # Closing actions
-            mininet_process.process.stdin.close()
-            mininet_process.read_stderr("Done")
-
-        # Write output to log file     
-        output_path = f"logs/{start_time_str}_output -t {iperf_time} -P {parallel_connections} {trial_runs} trials.log"
-        with open(output_path, 'w') as output_file:
-            output_file.write("Baseline Output:\n")
-            logging.info("Baseline Output:")
-            for i in baseline_output:
-                output_file.write(i + "\n")
-                logging.info(i)
-            output_file.write("Failover Output:\n")
-            logging.info("Failover Output:")
-            for i in failover_output:
-                output_file.write(i + "\n")
-                logging.info(i)
-
-        elapsed_time = (datetime.datetime.now() - start_time)
-        logging.info(f"{test_name} test done, elapsed time: {elapsed_time}")
-        print(f"Elapsed time: {elapsed_time}")
-
-        return (baseline_output, failover_output) 
-
-    except Exception as e:
-        logging.error(f"Error during {test_name} test: {e}")
-
 def bandwidth_control_test(testcase, topo="BW"):
     test_name = "Bandwidth Control"
     testcase_success = False
@@ -691,8 +341,6 @@ def bandwidth_control_test(testcase, topo="BW"):
             logging.error(f"Invalid testcase: {testcase}")
             return
 
-
-
         # Closing actions
         mininet_process.process.stdin.close()
         remaining_output = mininet_process.read_stderr("Done")
@@ -704,76 +352,64 @@ def bandwidth_control_test(testcase, topo="BW"):
 
     logging.info(f"{test_name} test done")
 
-def prototype(testcase, topo="TC"):
-    test_name = "Fault Tolerance prototype"
-    testcase_success = False
-    logging.info(f"Starting {test_name} test...")
-    
-    try:
-        mininet_process = MininetProcess(topo)
-        mininet_process.start_mininet()
-
-        baseline_throughput = None
-        # logging.info("Baseline iperf output")
-        mininet_process.read_logfile(f"iperf_baseline_{testcase}.log")
-        
-        baseline_output = mininet_process.read_logfile(f"iperf_baseline_{testcase}.log", last_line=True)
-        # mininet_process.read_logfile("iperf_failover.log")
-
-        # baseline_output = mininet_process.read_logfile("iperf_failover.log", last_line=True)
-        print(baseline_output)
-
-    except Exception as e:
-        logging.error(f"Error during {test_name} test: {e}")
-
-    logging.info(f"{test_name} test done")
-
 def test_selection():
-    user_input = input("Would you like to run a unit test or experiment?:\n"
-                       "1. Unit Test\n"
-                       "2. Experiment\n"
-                       "3. All\n"
-                       "4. Multi TCP\n"
+    user_input = input("Select a test category:\n"
+                       "1. Basic Tests (Pingall, Intents, Path)\n"
+                       "2. Link Automation (Basic Link Automation, Dynamic Path Automation, Fault Tolerance)\n"
+                       "3. Run All Tests (excludes Fault Tolerance test)\n"
+                    #    "4. Multi TCP\n"
                        "Enter: ")
     if user_input == "1":
-        unit_test = input("Which unit test would you like to run?:\n1. Pingall\n2. Intent Functions\n3. Path Functions\nEnter: ")
+        unit_test = input("Which basic functional test would you like to run?:\n1. Pingall\n2. Intent Functions\n3. Path Functions\n4. All\nEnter: ")
         if unit_test == "1":
             pingall_test()
         elif unit_test == "2":
             intent_functions_test(topo="DFGW")
         elif unit_test == "3":
             paths_functions_test(topo="DFGW")
+        elif unit_test == "4":
+            pingall_test()
+            intent_functions_test(topo="DFGW")
+            paths_functions_test(topo="DFGW")
+        else:
+            print("Invalid testcase number. Exiting...")
+            logging.error("Invalid testcase number. Exiting...")
     elif user_input == "2":
-        experiment = input("Which experiment would you like to run?:\n1. Basic Link Automation\n2. Dynamic Path Automation\n3. Fault Tolerance\n4. All\nEnter: ")
+        experiment = input("Which experiment would you like to run?:\n1. Basic Link Automation\n2. Dynamic Path Automation\n3. Fault Tolerance\nEnter: ")
         if experiment == "1":
             basic_link_auto_test(topo="DFGW")
         elif experiment == "2":
-            testcase = input("Which testcase would you like to run?:\n1. Reactive Forwarding\n2. Host Intents\nEnter: ")
+            testcase = input("Which testcase would you like to run?:\n1. Reactive Forwarding\n2. Host Intents\n3. Both\nEnter: ")
             if testcase == "1":
                 dynamic_paths_test(testcase="fwd", topo="DFGW")
             elif testcase == "2":
+                dynamic_paths_test(testcase="intent", topo="DFGW")
+            elif testcase == "3":
+                dynamic_paths_test(testcase="fwd", topo="DFGW")
                 dynamic_paths_test(testcase="intent", topo="DFGW")
             else:
                 print("Invalid testcase number. Exiting...")
                 logging.error("Invalid testcase number. Exiting...")
         elif experiment == "3":
-            testcase = input("Which testcase would you like to run?:\n1. Reactive Forwarding\n2. Host Intents\n3. All\nEnter: ")
-            if testcase == "1":
-                fault_tolerance_test(testcase="fwd", topo="TC")
-            elif testcase == "2":
-                fault_tolerance_test(testcase="intent", topo="TC")
-            elif testcase == "3":
-                fault_tolerance_test(testcase="fwd", topo="TC")
-                fault_tolerance_test(testcase="intent", topo="TC")
-            else:
-                print("Invalid testcase number. Exiting...")
-                logging.error("Invalid testcase number. Exiting...")
-        elif experiment == "4":
-            basic_link_auto_test(topo="DFGW")
-            dynamic_paths_test(testcase="fwd", topo="DFGW")
-            dynamic_paths_test(testcase="intent", topo="DFGW")
-            fault_tolerance_test(testcase="fwd", topo="TC")
-            fault_tolerance_test(testcase="intent", topo="TC")
+            print("This testcase has been disabled due to the time it takes to run (20 trials of 20s iperf tests)\nRun fault_tolerance_test.py instead.")
+            logging.error("This testcase has been disabled due to the time it takes to run (20 trials of 20s iperf tests) Run fault_tolerance_test.py instead.")
+            # testcase = input("Which testcase would you like to run?\n(These will take awhile to run since 20 trials are conducted to generate statistics, you may want to view log files simultaneously):\n1. Reactive Forwarding\n2. Host Intents\n3. Both\nEnter: ")
+            # if testcase == "1":
+            #     fault_tolerance_test(testcase="fwd", topo="TC")
+            # elif testcase == "2":
+            #     fault_tolerance_test(testcase="intent", topo="TC")
+            # elif testcase == "3":
+            #     fault_tolerance_test(testcase="fwd", topo="TC")
+            #     fault_tolerance_test(testcase="intent", topo="TC")
+            # else:
+            #     print("Invalid testcase number. Exiting...")
+            #     logging.error("Invalid testcase number. Exiting...")
+        # elif experiment == "4":
+        #     basic_link_auto_test(topo="DFGW")
+        #     dynamic_paths_test(testcase="fwd", topo="DFGW")
+        #     dynamic_paths_test(testcase="intent", topo="DFGW")
+            # fault_tolerance_test(testcase="fwd", topo="TC")
+            # fault_tolerance_test(testcase="intent", topo="TC")
     elif user_input == "3":
         pingall_test()
         paths_functions_test("DFGW")
@@ -781,10 +417,10 @@ def test_selection():
         basic_link_auto_test("DFGW")
         dynamic_paths_test(testcase="fwd", topo="DFGW")
         dynamic_paths_test(testcase="intent", topo="DFGW")
-        fault_tolerance_test(testcase="fwd", topo="DFGW")
-        fault_tolerance_test(testcase="intent", topo="DFGW")
-    elif user_input == "4":
-        iperf3_test("fwd", topo="TC", iperf_time=20, trial_runs=20, parallel_connections=3)
+        # fault_tolerance_test(testcase="fwd", topo="DFGW")
+        # fault_tolerance_test(testcase="intent", topo="DFGW")
+    # elif user_input == "4":
+        # iperf3_test("fwd", topo="TC", iperf_time=20, trial_runs=20, parallel_connections=3)
     else:
         print("No test selected. Exiting...")
         logging.error("No test selected. Exiting...")
